requirements/base.txt
django==3.1.7  
django-environ==0.4.5  
django-model-utils==4.1.1  
django-allauth==0.44.0  

argon2-cffi==19.2.0  
coreapi==2.3.3 
django-cors-headers==3.2.1
gunicorn==20.0.4
regex==2020.4.4
pyyaml==5.4.1
argh==0.26.2 
watchdog==0.10.2
celery==4.4.7
django-celery-results==1.2.1
django-celery-beat==2.2.0
djangorestframework-simplejwt==4.4.0
django-filter==2.4.0